never create real unnecessary  markup in the document_tags -> frontend method.
for example we had a lot of pain if we wanted to create pdf's using a html2pdf converter because of this div tag.
so if someone wants to have such an html element, he can create it in the view. pimcore specific markup must not be displayed in the content!