Explicitly set file mode for optimized images
For me, jpegoptim set the file mode of optimized images to 0600 which made them unreadable for our webserver process. I've added chmods in the same way it is done for "regular" thumbnails to prevent this from happening. Though i have not tested the other optimizers i suspect the same problem might occur, so i added chmods for them as wenn, just in case