Fix ApplicationLogger logging source showing inconsistent results
Depending on the way the logger is called it showed different results. A
call to `log()` showed the real source, while a call to `info()` showed
`info()` as the log source. Additionally, when called from a context
outside a class (plain PHP file, function), the source and the caller
line was wrong.

Fixes #562