suppress class_alias warnings legacy classmapping
When you have CLI scripts that call Pimcore::initPlugins() for some
reason, the legacy classmapping will run the class aliasses again
causing the output of a truckload of warnings. Here the warnings are
just supressed because its not really a problem and there does not seem
to be a clean solution to check if the class alias is already there.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>