Call update video panel in another way - see #1121
`this.getBody().update(content);`
seemed to mess up the dom. It removed the inner div's of the classes 'x-autocontainer-outerCt' and 'x-autocontainer-innerCt', which caused Ext.js to crash.
If you update the html of the video panel component like this:
`this.component.update(content);`
It will keep the dom elements in place.