Fall back to parent object ID if parentId is not set but we have a parent object #1150
If one object is set as the parent of another object, but the parent
object was not saved at the time of the setParent(), the child object
will never get the real parent ID after the parent was saved. This reads
the parent ID directly from the parent if a parent is set but its ID is
not.