Add extension point to initialize context specific state through context initializers
Initializers will be called as soon as the context listener resolved a
context and are added to a DelegatingContextInitializer through the
pimcore.context_initializer DI tag.

A context initializer can decide if it should run for a context and has
access to request and request type to further decide what to initialize.