[SECURITY] Make password encoders which need a user context way more configurable
As password scenarios handling pimcore users or objects with a Password
field definitions need access to the user instance, we need to built a
dedicated encoder per user instance.

This change adds the possibility to define an encoder factory per class
which is responsible for creating the encoder and optionally building a
dedicated one per user object (see UserAwareEncoderFactory).

The provided classes can be configured via DI and should fit a lot of
default scenarios. The class/factory mapping expects the factory to be
defined on the container and the mapping configured in
pimcore.security.encoder_factories.