Fix Pimcore::inDebugMode() returning false if debug mode is configured to IP
Tool::getClientIp() does not have access to a request at the time the
kernel is built. Now the request is set as static property on Tool until
the kernel is built. To make this possbile, initialization logic needs
to be split up into bootstrapping (autoloader, constants) and building
the kernel.