[ECOMMERCE-CONFIG][CHECKOUT-MANAGER] Define checkout manager config tree and build checkout managers through a factory service
Checkout managers have an additional level of categorization as there
can be multiple named checkout managers which in turn define multiple
tenants. As they are cart specific (which is defined at runtime), a
dedicated checkout manager factory and commit order processor are built
for every name+tenant combination. When requesting a checkout manager
for a given name/tenant, the matching factory will be loaded and used
to build the requested manager.