[INSTALLER] Remove implicitely added parent directories from list of files
E.g. files to add:

- app/config/pimcore/*
- app/config/pimcore/ecommerce/*

In this case we can't copy/symlink the app/config/pimcore/ecommerce
directory directly (which would be included by the first line , as we
need to handle the files inside the directory one-by-one. This change
removes the app/config/pimcore/ecommerce from the list of files. The
directory is still implicitely added, but not as single copy/symlink on
the directory level, but for every file inside the directory.