[INSTALLER] Make sure container isn't built before config files were created by not building Model\Tool\Setup befor new kernel is booted
Timing issue which can't be reproduced in a deterministic way. As
building Model\Tool\Setup before building the new AppKernel, a container
is already built with the config files/values existing at that moment
(before copying the new config) and creating a system.php file with the
default values (empty DB credentials). Sometimes this leads to those
credentials still being cached in the container after the installer was
executed.

This moves the config file creation logic to a new class which can be
instantiated without relying on a model class/dao and without
prematurely building the container instance without the new config
already in place.