[MIGRATIONS] Make migration sets configurable via config
This adds the possibility to define migrations sets which do not belong
to a bundle via configuration (as the default app configuration set).

This introduces a small BC break as the constructor signature and the
buildDefaultMigrationSets() method in ConfigurationFactory changed, but
makes the migrations setup more usable in the long run.