setFrom() does not accept array of mail addresses
The `setFrom($email, $name = null)` function of `\Pimcore\Mail` is not working if an array gets passed as `$email` parameter, which is not compatible with the `addFrom()` function of the extended class `\Swift_Mime_SimpleMessage`. This produces an error in line 543 `Warning: preg_match() expects parameter 2 to be string, array given`.