[TARGETING] Use storage data instead of Piwik to handle visited-n-pages before
The condition reads the current page count from storage instead of using
Piwik's API to fetch visit date. After matching it emits an event which
instructs the VisitedPagesCountListener to increment the count after all
matching proceeded. To make sure the count is incremented only once per
request, the incrementing logic is encapsulated in a service which is
used from the condition to query data and from the event listener to
increment the count.

Potential improvement: Make the storage scope choosable when configuring
the condition.