Fix exceptions in lost password and login as different user actions when user has no password set
As the generated token is a value which is encrypted with the current
password hash we would need to generate a random password before being
able to generate a token for a user without password. As generating a
password implicitely can lead to security issues we disallow
lostpassword/login as different user for users without a password.