Remove hardcoded environments from multiple locations and add them to a configurable EnvironmentConfig class
EnvironmentConfig can either be reconfigured using setters or completely
replaced by implementing a custom config class. Environments to register
in development are still hardcoded on the kernel, but extracted into a
method which can be overridden. InstallerKernel still uses hardcoded
environments, but that should be fine for the purpose of the installer.