Make placeholder view helpers using the container (e.g. headTitle) sub-request aware
When rendering many documents in the same request (e.g. as the SEO
document editor does), the shared state of the same placeholder
container affects different documents leading for example to titles
containing all parts of previously rendered documents (getting longer
and longer).

This change introduces pre and post render events into the document
renderer which are consumed from the placeholder registry to initialize
a new, empty state when a new document is rendered.