Fix cover transformation when focal point near edge
If an image has a focal point too close to the right edge or the bottom, the cover transformation will end up cropping the image too much, leaving the resulting thumbnail narrower or shorter than it's supposed to be. This fix puts a ceiling on the cropX and cropY values in order to preserve the desired height and width of the thumbnail.