Fix Voucher FIeldcollection charactertype and length settings
As is currently the case the settings for this field allows the user to leave it as-is, causing `Pattern::generateCode()` to throw a fatal error and fail. The same goes for Length's `minValue` where a negative (why anyone would ever have a negative length I don't know) will cause a fatal error to be thrown. The length value needs to be `prefix+1` and prefix can be empty, so a value of  `1` should be a sane and usable default.