Key of assoc array is lost when using the webservice
For example, when accessing a document via the webservice in the browser directly, you will see the JSON that will be sent to any bundles using the webservice (e.g. RestImporter). If there are any elements like an Areablock the indices will contain all information as non-associative array. This also means that importing data (in case of the RestImporter) is not possible.