Error handling and logging
This is a proposal/food for thought... For context, we've been trying to run Pimcore in a set of Docker containers.

# is_integer vs defined

Especially when you use environment variables to configure this setting, it's never ever going to be an integer as the code to "resolve constants" has no context/domain knowledge of the setting, doesn't cast the constant and therefor it's of type `string` and the `is_integer` is always false. You could probably use `is_numeric()` or so. But I haven't tested it.

# error_log

I changed this as well as it doesn't work when you e.g. set the `error_log` to `/proc/self/fd/2` to catch logs on `stdout` or `stderr`.

I would assume that when people set this environment variable (or the constant) they know what's going on and checking for writable is not necessary. The `is_writable` is also a hefty performance penalty each time this code is executed too. Removing this definitely fixes "my" problem, but I am also sure there are other use cases and maybe someone else has a better idea.