[ClassificationStorre] Error validating mandatory fields
Error message: Call to a member function getMessage() on array
Caused by a series of events:
In ClassificationStore::checkValidity() the ValitionException sub items are set
with an array of error data instead of nested ValidationExceptions.
In Concrete::update() the ValidationException is rethrown when inheritance is
allowed.
In AdminExceptionListener::onKernelException() the method
recursiveAddValidationExceptionSubItems() is called, which in turn causes the
error because it expects the sub items to be ValidationException objects and
tries to call getMessage() in the foreach loop.
Fixed by changing sub items to ValidationException objects in
ClassificationStore::checkValidity()