A possibility to define environment specific constants
The number of files to be changed in a deployment should be as small as possible. As the constants file is loaded before the environment i detected, we should have a possibility to define environment specific constants after the environment has been defined. 

# Use case
Disabling Amazon S3 on local (defining S3 variables only for staging environment).