change[Thumbnails]: make auto clipping configurable (#6950)
* fix [Imagick]: add compatibility fix for Imagick 3.4.4 and ImageMagick 7

regression from https://github.com/pimcore/pimcore/issues/5505

* change[Thumbnails]: make auto clipping configurable

https://github.com/pimcore/pimcore/issues/5505 introduced auto clipping for images
with embedded clipping path.

This is a pretty wonky feature and often broken through several ImageMagick releases. As well
this is a high cpu intensive task. People should be carful to use it.

hence we make it globaly configurable and enable it by default

* Update doc/Development_Documentation/04_Assets/03_Working_with_Thumbnails/01_Image_Thumbnails.md

Co-authored-by: Bernhard Rusch <brusch@users.noreply.github.com>