[DataObject][Inheritance] Fix inheritance when using default values (#7033)
DefaultValueTrait::handleDefaultValue() determines the parent's value
and decides whether or not to use the default value. This worked for
the first child (1st level of inheritance) but not for the next child
(2nd level), as GetInheritedValues is disabled in save() context.

The result is unwanted persisting of the default value to store table
for the 2nd child as thus effectively breaking inheritance.

* object
|-- variant 1: OK
    |-- variant 2: NOT OK
        |-- variant 3: OK
            |-- variant 4: NOT OK
                ...

This fix enables inheritance while determining the value from the parent
and thus make the right decision whether to use the default or not.