Update NotificationEmailService.php
Allows notification email to link correctly to items other than objects. Changing a document or asset and submitting a workflow against the item produces an email with a link to an object "...object_ID_object..."

This code selects the appropriate type from the currently available types or just defaults to object.