Fix race condition in workflow transition saves
This fixes a problem where a save is overwritten with an older save when you save an object in a event triggered by a transition. After saving the object, another save is triggered in the WorkflowManager that puts the object in an older state again undoing everything we did in the transistion event itself. The save itself seems not needed since both notes and transitions are being updated correctly even without this save