fix: shift timezone on date field
Error: Depending on the timezone of the browser and the timezone of the
Pimcore server the stored date may be either correct, one day earlier or
one day later than the date entered by the user.

Caused by: The date is selected in the browser using the user's timezone
and sent to the server as timestamp. For example if the user is in CEST
and selects "2021-08-03" then the timestamp for "2021-08-03 00:00:00
CEST" is sent to the server. But if the server is configured to use the
New York timezone, it converts that timestamp to "2021-08-02 18:00:00
EDT" and thus stores "2021-08-02" (if DATE MySQL column is used).
Similar issues occurs for the TIMESTAMP MySQL column.

Resolved by: Shifting timezone on the entered date when receiving data
from or sending data to Pimcore.