fix: advanced relation metadata double click
When double clicking metadata fields of an advanced many to many relation row
the related data object is opened. This is very annoying, as it requires two
clicks to open metadata select dropdowns or focus the metadata input field.
Resolved by removing the double click event for the advanced relation fields.
An open icon is already available to open the record.
Resolves #10051