Website settings fetch changed
For multidomain environments website settings are not selected correctly. 
In my case a custom website setting "contactFormEmailRecipient" is set for main site and a sub site. The`$data` array holds two results and for the sub site `$data[0]` is taken. This is not correct in my case. With this fix, `$data` holds only one result and everything is fine.