Update select.js
The selected value of a ComboBox breaks when the key is Html. It seems to be caused by the sencha Library, reproducible in the sencha sandbox as well.
The changes resolve the problem with using a template for the options ("tpl") and a template for the selected value/key ("displayTpl"), and the problem causing dispalyfield is gone.