[Perspective] Get rid of not allowed perspective warning when not requesting a perspective
On the default installation every time reloading the admin ui with the default GET-parameter "&perspective=", a WARNING "User pimcore is not allowed requested perspective ." is logged. But as seen a few lines before, this is not stated as a request for a now perspective. "Perspective for user {user} was not requested." as debug log makes sense.