[Perspective] Omit "not allowed perspective"-warning with &perspective= as GET-Param
With the default installation every time reloading the admin UI with the default GET-parameter "&perspective=", a WARNING "User pimcore is not allowed requested perspective ." is logged. But as seen a few lines above, this is not stated as a request for a new perspective. "Perspective for user {user} was not requested." as debug log makes sense here which is logged after this change.

Additionally, fixed wording in a warning log-message.