Reset fallback document after assigning it
A problem occurs in some special cases, i.e. when a cli script or some long running processes perform subsequent requests like so:

```
$kernel   = Kernel::get();
$request  = Request::create($somePath, 'GET');
$response = $kernel->handle($request);
```

The problem is that in the above use case the document once assigned in line 119 will be used in subsequent calls. That will overwrite an already correctly set document in the DocumentResolver in line 159.

When the fallback document is set to `null` afterwards, that won't happen.