[Admin] Fix displaying multiselect with options provider and render type "list"
If a multiselect (render type "list") with an options provider returns numeric IDs
(as given as an example in the documentation) the multiselect element isn't correctly
rendered in the edit layout: The set items are not selected.

This fix forces the "id" field of the store to be handled as a string, exactly as a
static definition of items would be.