fix: dynamic multiselect options not loaded in grid
When using an options provider to generate options depending on a value from
the data object the options were not loaded in the grid, because the data
object is not available when retrieving the general grid configuration.
It does work for the regular select, because it has a separate grid data
implementation for the grid-proxy setup.
Fixed by adding the same implementation for the multiselect field and updating
the JavaScript file accordingly.
Resolves issue #12029