Update grid.js
When trying to add a column to the grid I get an error `fieldConfig.layout is undefined`. I tried to debug the code and realized that that situation is as follows:

// Don't add batch options if field has dynamic options
console.log('hasOwnProperty: ' + fieldConfig.hasOwnProperty('layout')); // => true
console.log('undefined: ' + (fieldConfig.layout === undefined));    // => true

So, `hasOwnProperty` is true but `layout` === undefined is true too. The suggested solution is to add a condition to check `layout` for undefined.