[Improvement] edit not editable advanced relation columns
Allow editing columns of advanced many to many (object) relation fields when
not editable is enabled by providing a new option:
Allow editing columns when not editable
This allows the columns to be used as configuration, while the user is unable
to add or (re)move rows. Rows can be added by administrators or via event
listeners, for example.